requirements.txt
Cerberus==1.3.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
haversine==2.1.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
Werkzeug==0.15.5
